Fix eye gradients and extract function mask analysis
